Revert "Fix position encoding for non-equip items"
This reverts commit c1f678637d73b3d2cf6d867f435d8d61481db9c1.

+++ src/main/java/tools/PacketCreator.java
@@ -418,7 +418,7 @@ public class PacketCreator {
                     p.writeShort(pos);
                 }
             } else {
-                p.writeShort(pos);
+                p.writeByte(pos);
             }
         }
         p.writeByte(itemType);
